add Redis to fix deadlock problem

+++ src/main/java/ticket/test/ticketing/TicketingService.java
@@ -1,10 +1,10 @@
 package ticket.test.ticketing;
 
-import jakarta.persistence.LockModeType;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.redisson.api.RLock;
+import org.redisson.api.RedissonClient;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.jpa.repository.Lock;
 import org.springframework.stereotype.Service;
 
 import org.springframework.transaction.annotation.Transactional;
@@ -13,26 +13,49 @@ import ticket.test.ticketing.db.TicketRepository;
 
 import java.util.Date;
 import java.util.Optional;
+import java.util.concurrent.TimeUnit;
 
 @RequiredArgsConstructor
 @Slf4j
 @Service
+@Transactional()
 public class TicketingService {
 
     @Autowired
     private final TicketRepository ticketRepository;
 
+    private final RedissonClient redissonClient;
+
     /*
     *  티켓을 예약한다.
     */
-    @Transactional()
+
     public TicketingResponseDto getTicket(TicketingRequestDto request) {
+        TicketingResponseDto result;
+        RLock lock = redissonClient.getLock("showTicket_" + request.getShowId());
+        try {
+            if(lock.tryLock(10000, 3000, TimeUnit.MILLISECONDS)) {
+                result = saveTicket(request);
+            } else {
+                throw new RuntimeException("Unabled to acquire lock");
+            }
+        } catch (InterruptedException e) {
+            throw new RuntimeException(e);
+        } finally {
+            if(lock.isLocked() && lock.isHeldByCurrentThread()) {
+                lock.unlock();
+            }
+        }
+        return result;
+    }
+
+    private TicketingResponseDto saveTicket(TicketingRequestDto request) {
         Ticket duplicate = ticketRepository.findByUserIdAndShowId(request.getUserId(), request.getShowId());
         if(duplicate != null) {
             return new TicketingResponseDto("이미 예약된 공연입니다.", null);
         }
 
-        Optional<Ticket> savedSeat = ticketRepository.findByShowIdAndSeatForUpdate(request.getShowId(), request.getSeat());
+        Optional<Ticket> savedSeat = ticketRepository.findByShowIdAndSeat(request.getShowId(), request.getSeat());
         if(savedSeat.isPresent()) {
             return new TicketingResponseDto("이미 선점된 자리입니다.", null);
         }
@@ -41,7 +64,6 @@ public class TicketingService {
                 request.getTicketId(), request.getUserId(), request.getShowId(), request.getSeat(), new Date()
         );
         ticketRepository.saveAndFlush(ticket);
-
         return new TicketingResponseDto("예약이 완료되었습니다.", ticket);
     }
 
@@ -50,11 +72,25 @@ public class TicketingService {
      */
     @Transactional()
     public TicketingResponseDto cancelTicket(TicketingRequestDto request) {
-        ticketRepository.findByTicketIdForUpdate(request.getTicketId()).orElseThrow(() -> {
-            throw new IllegalArgumentException("해당 ID가 존재하지 않습니다.");
-        });
+        TicketingResponseDto result;
+        RLock lock = redissonClient.getLock("cancelTicket_" + request.getTicketId());
+        try {
+            if(lock.tryLock(10000, 3000, TimeUnit.MILLISECONDS)) {
+                ticketRepository.findByTicketIdForUpdate(request.getTicketId()).orElseThrow(() -> {
+                    throw new IllegalArgumentException("해당 ID가 존재하지 않습니다.");
+                });
+                ticketRepository.deleteByTicketId(request.getTicketId());
+            } else {
+                throw new RuntimeException("Unabled to acquire lock");
+            }
+        } catch (InterruptedException e) {
+            throw new RuntimeException(e);
+        } finally {
+            if(lock.isLocked() && lock.isHeldByCurrentThread()) {
+                lock.unlock();
+            }
+        }
 
-        ticketRepository.deleteByTicketId(request.getTicketId());
         return new TicketingResponseDto("예약 취소가 완료되었습니다.", null);
     }
